+++ src/utils/dynamicDescription.ts
@@ -13,7 +13,7 @@ export const preferPaged = (text: UnknownText) => {
 export const createDynamicText = (text: UnknownText): TextFunction => {
   if (typeof text === "string" || text instanceof Text || text instanceof ManagedText) {
     return () => text;
-  } else if (Array.isArray(text) && typeof text[0] === "string") {
+  } else if (Array.isArray(text)) {
     const sequence = new CyclicText(...text);
     return () => sequence;
   } else if (typeof text === "function") {
@@ -21,6 +21,6 @@ export const createDynamicText = (text: UnknownText): TextFunction => {
   } else if (typeof text === "undefined") {
     return (item: string) => `There's nothing noteworthy about the ${item}.`;
   } else {
-    throw Error("Text must be a string, string array, Text, ManagedText or a function");
+    throw Error("Text must be a string, Text, ManagedText, function or an array of the above.");
   }
 };